requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
tensorflow>=2.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
mysql-connector-python>=8.0.0
xgboost>=1.7.0
huggingface_hub>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0